backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
playwright
openai
python-dotenv
sqlalchemy
PyJWT[crypto]
alembic
requests
